enable_if and is_integral implementation

+++ iterator.hpp
@@ -12,7 +12,7 @@ namespace ft
 
 namespace ft
 {
-	template<class Category, class T, class Distance = long, class Pointer = T*, class Reference = T& >
+	template<class Category, class T, class Distance = ptrdiff_t, class Pointer = T*, class Reference = T& >
 	struct iterator
 	{
 		typedef typename	Category	iterator_category;
